Update review.installation for existing manifest flow
- Required refactoring useExistingManifest utilities to their own
mixin, if the page was refreshed from review.installation, the
subscriptions model is not available. review.installation must load that
data itself if required.
