+++ fusor-ember-cli/app/routes/review/installation.js
@@ -1,7 +1,8 @@
 import Ember from 'ember';
 import request from 'ic-ajax';
+import NeedsExistingManifestHelpers from '../../mixins/needs-existing-manifest-helpers';
 
-export default Ember.Route.extend({
+export default Ember.Route.extend(NeedsExistingManifestHelpers, {
 
   beforeModel() {
     // Ensure the models have been persisted so that we're validating/syncing up to date data.
@@ -18,12 +19,39 @@ export default Ember.Route.extend({
     return Ember.RSVP.hash(promises);
   },
 
-  setupController(controller, model) {
+  model() {
+    const reviewModel = this.modelFor('review');
+    const subModel = this.modelFor('subscriptions');
+    let modelHash = {reviewModel};
+    if(subModel) {
+      // Use subscriptions model if the loading has already been done
+      modelHash.useExistingManifest = subModel.useExistingManifest;
+      if(subModel.subscriptions) {
+        modelHash.subscriptions = subModel.subscriptions;
+      }
+      return modelHash;
+    } else {
+      // subscriptions model isn't available, maybe because of a page refresh
+      // Need to load this data independently
+      return this.shouldUseExistingManifest().then(useExistingManifest => {
+        modelHash.useExistingManifest = useExistingManifest;
+
+        if(useExistingManifest) {
+          modelHash.subscriptions = this.loadSubscriptions();
+        }
+
+        return Ember.RSVP.hash(modelHash);
+      });
+    }
+  },
+
+  setupController(controller, modelHash) {
+    const model = modelHash.reviewModel;
     controller.set('model', model);
     controller.set('showErrorMessage', false);
+    controller.set('useExistingManifest', modelHash.useExistingManifest);
     if (model.get('deploy_rhev')) {
       this.store.findAll('hostgroup').then(function(results) {
-        console.log(results);
         var fusorBaseHostgroup = results.filterBy('name', 'Fusor Base').get('firstObject');
         var fusorBaseDomain = fusorBaseHostgroup.get('domain.name');
         controller.set('engineDomain', fusorBaseDomain);
@@ -31,7 +59,11 @@ export default Ember.Route.extend({
       });
     }
 
-    if (model.get('is_disconnected')) {
+    const subModel = this.controllerFor('subscriptions');
+    if(modelHash.useExistingManifest) {
+      controller.set('useExistingManifest', true);
+      controller.set('reviewSubscriptions', modelHash.subscriptions);
+    } else if (model.get('is_disconnected')) {
       controller.set('reviewSubscriptions', this.modelFor('subscriptions/review-subscriptions'));
     } else {
       var reviewSubscriptions = model.get('subscriptions').filter(function(sub) {
@@ -42,7 +74,7 @@ export default Ember.Route.extend({
       }, false); // initial val
       controller.set('reviewSubscriptions', reviewSubscriptions);
       controller.set('hasSubscriptionsToAttach', hasSubs);
-      controller.set('hasSessionPortal', Ember.isPresent(this.modelFor('subscriptions').sessionPortal));
+      controller.set('hasSessionPortal', subModel ? subModel.sessionPortal : null);
       controller.set('hasSubscriptionPools', Ember.isPresent(this.controllerFor('subscriptions/select-subscriptions').get('subscriptionPools')));
     }
